Make Collection cache keys after the first lookup.
I didn't want to do this, but searching for a Node key several thousand
times when creating the links (graph structure) is still too slow when
allowing keys of documents to change.

Instead, I added Collection#refresh which will return a new Collection
without the key cache, and will use the latest version of the keys the
next time you call "find".