Completely remove Slot documents
The Slot class remains, mostly to make it easier to read the attributes
from the node; so that you aren't having to mess about with attributes
on a hash, but can instead call Node#in_slots or Node#out_slots.

The Node "efficiency" attribute has been renamed to "output". e.g.

    - output.heat = 0.47
    - output.electricity = 0.43

... and an "input" attribute, which corresponds to "in" slots has also
been added.