GraphBuilder supports flexible links and a super-source
Adds support for setting links to be type=:flexible which are calculated
exclusively from the left ("child") node.

Nodes which receive energy from an external source (in another
sub-graph) are now connected to a SUPER_SOURCE. This makes up for the
deficit in supply of steam/hot-water in the agriculture graph.