ProductionMode is now created with a hash
This is instead of referring to some pre-defined location in ETSource,
so that we can more easily test production mode objects in ETEngine.
But, more importantly, allows us to "lazy-compile" the graph when needed
in ETEngine, and save the static file to its ./tmp direction instead of
it having to exist in ETSource.