fail -> Kernel.raise
TravisCI's Ubuntu image was failing as "fail(...)" was returning a
Rubel::Message instead of raising an exception. Made this call more
precise by calling the version on Kernel directly.

This is probably because Rubel::Base inherits from BasicObject, but that
doesn't quite explain why it passes on a local copy of Ubuntu nor on
OS X. Perhaps some quirk of the container?