Add "scaling_exempt" attribute to Node
Permits a node to be excluded from the graph scaling when creating a
derived dataset, or refreshing the graph.yml.

Changes GraphScaler to receive the Graph, rather than the hash export of
a graph, mainly to provide easy access to the model (Node). The result
is then passed to the EssentialExporter to be dumped to YAML.

Closes quintel/etsource#1452