Calculating derived datasets should use itself
When calculating a derived dataset, we were passing in the parent as the
source - presumably because the derived dataset is lacking some files
such as the energy balance, primary production, etc.

However, this means that `AREA` lookups would return values from the
parent, rather than the derived dataset. This was most obvious where
national-sized electricity interconnectors were given to small regions.

Ref quintel/etsource#1757
Ref quintel/etmodel#3114
Ref quintel/etmodel#3385