Fix ProductionManager having wrong dir or ext
Fixes that ProductionManager would have an incorrect directory_name or
extension_name when the class being managed has a custom value for
either or both.

This would result in the manager finding no objects (on the file
system) to manage.

`lookup_map` is now filtered to ignore any objects on the file system
for which data wasn't provided.