Add PathResolver and path resolution to Dataset
PathResolver may be given one or two paths to directories and will look
for files in the order the directories are provided.

For example, this is useful for derived datasets where we want to omit
files which are identical to those in the parent dataset, while adding
only those files which need to be customised. PathResolver can be given
both the dataset and parent dataset path, and given a relative path to a
file, resolve the path the file in either dataset.