Read GPG encrypted energy balances
Previously it was the responsibility of the application to decrypt
energy balance files prior to reading from ETSource. This is annoying,
and prone to mistakes.

Atlas will now look for an `ETSOURCE_PASSWORD` environment variable, or
a .password file in the ETSource data directory, containing the password
to decrypt GPG energy balances.